+++ jobs/Backend/Task/Configuration/CzechBankSettings.cs
@@ -4,5 +4,7 @@
     {
         public string DailyRatesUrl { get; set; }
         public string OtherCurrencyRatesUrl { get; set; }
+        public int TimeoutSeconds { get; set; } = 10;
+        public int RetryCount { get; set; } = 3;
     }
 }