+++ jobs/Backend/Task/Parsers/IExchangeRateParser.cs
@@ -1,0 +1,10 @@
+﻿using System.Collections.Generic;
+using ExchangeRateUpdater.Models;
+
+namespace ExchangeRateUpdater.Parsers
+{
+    public interface IExchangeRateParser
+    {
+        IEnumerable<ExchangeRate> Parse(string content, IEnumerable<Currency> filter, Currency source);
+    }
+}