+++ jobs/Backend/Task/CHANGELOG.md
@@ -8,8 +8,21 @@ and this project adheres to [Semantic Versioning](https://semver.org).
 ## [Unreleased]
 
 ### Added
+- 2 HttpClients which fetch the Czech National Bank data sources
+- Parser for parsing data from the http call
+- Integration Test testing entire flow with mocked Http Responses
+- The `CHANGELOG` file itself
+- `Deployment.md` and `Dockerfile` for deployment steps
 
 ### Fixed
+- Single Responsibility Principle was being violated by the `ExchangeRateProviderService`
+- Moved the `ExchangeRateProvider` to Services folder keeping root clean
+- Unit tests which were failing from the new `ExchangeRateProvider` implementation
+- csproj organisation and using .net 9 packages on a project with target of .net 6
+- `appsettings` to folow the removal of `HttpServiceSettings`
+
+### Removed
+- `HttpServiceSettings`; moved into the `CzechBankSettings`
 
 ---